+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/GenerateParserTask.kt
@@ -47,31 +47,57 @@ abstract class GenerateParserTask : JavaExec() {
     abstract val targetRootOutputDir: DirectoryProperty
 
     /**
-     * Required.
      * The location of the generated parser class, relative to the [targetRootOutputDir].
+     * The value of this property is only used when [purgeOldFiles] is set to `true`.
+     * When this property is set, only the given file or directory is deleted before re-generating the parser.
      */
+    @Deprecated(
+        message = "The property is removed without replacement. " +
+                "This property is only relevant when the directory at `targetRootOutputDir` overlaps with other files. " +
+                "Note that overlapping task outputs are discouraged by Gradle and can cause issues with the build cache.",
+        level = DeprecationLevel.WARNING,
+    )
     @get:Input
+    @get:Optional
     abstract val pathToParser: Property<String>
 
     /**
-     * Required.
-     * The location of the generated PSI files, relative to the [targetRoot].
+     * The location of the generated PSI files, relative to the [targetRootOutputDir].
+     * The value of this property is only used when [purgeOldFiles] is set to `true`.
+     * When this property is set, only the given file or directory is deleted before re-generating the parser.
      */
+    @Deprecated(
+        message = "The property is removed without replacement. " +
+                "This property is only relevant when the directory at `targetRootOutputDir` overlaps with other files. " +
+                "Note that overlapping task outputs are discouraged by Gradle and can cause issues with the build cache.",
+        level = DeprecationLevel.WARNING,
+    )
     @get:Input
+    @get:Optional
     abstract val pathToPsiRoot: Property<String>
 
     /**
      * The output parser file computed from the [pathToParser] property.
      */
+    @Deprecated(
+        message = "The method will be removed together with `pathToParser`.",
+        replaceWith = ReplaceWith("targetRootOutputDir.file(pathToParser)"),
+        level = DeprecationLevel.WARNING,
+    )
     fun parserFile(): Provider<RegularFile> = targetRootOutputDir.file(pathToParser)
 
     /**
      * The output PSI directory computed from the [pathToPsiRoot] property.
      */
+    @Deprecated(
+        message = "The method will be removed together with `pathToPsiRoot`.",
+        replaceWith = ReplaceWith("targetRootOutputDir.dir(pathToPsiRoot)"),
+        level = DeprecationLevel.WARNING,
+    )
     fun psiDir(): Provider<Directory> = targetRootOutputDir.dir(pathToPsiRoot)
 
     /**
-     * Purge old files from the target directory before generating the lexer.
+     * Purge old files from the target directory before generating the parser.
      */
     @get:Input
     @get:Optional
@@ -79,10 +105,20 @@ abstract class GenerateParserTask : JavaExec() {
 
     @TaskAction
     override fun exec() {
+        if (pathToParser.isPresent && !pathToPsiRoot.isPresent) {
+            throw GradleException("'pathToParser' has a configured value, but 'pathToPsiRoot' has not. You must either remove or keep both properties.")
+        } else if (pathToPsiRoot.isPresent && !pathToParser.isPresent) {
+            throw GradleException("'pathToPsiRoot' has a configured value, but 'pathToParser' has not. You must either remove or keep both properties.")
+        }
         if (purgeOldFiles.orNull == true) {
-            targetRootOutputDir.get().asFile.apply {
-                resolve(pathToParser.get()).deleteRecursively()
-                resolve(pathToPsiRoot.get()).deleteRecursively()
+            if (!pathToParser.isPresent && !pathToPsiRoot.isPresent) {
+                targetRootOutputDir.get().asFile.deleteRecursively()
+            } else {
+                // Delete only the directories specified by `pathToParser` and `pathToPsiRoot` for backwards compatibility.
+                targetRootOutputDir.get().asFile.apply {
+                    resolve(pathToParser.get()).deleteRecursively()
+                    resolve(pathToPsiRoot.get()).deleteRecursively()
+                }
             }
         }
         ByteArrayOutputStream().use { os ->