+++ CHANGELOG.md
@@ -9,7 +9,9 @@
 ### Changed
 
 - Deprecate property `targetOutputDir` of `GenerateLexerTask`.
+- Deprecate properties `pathToParser` and `pathToPsiRoot` of `GenerateParserTask`.
 - Deprecate `targetFile(String)` and `targetFile(Provider)` of `GenerateLexerTask`.
+- Deprecate `parserFile()` and `psiDir()` of `GenerateParserTask`.
 
 ### Fixed